Clean up old scripts replace with xarray based ones for limited RAM
